requirements-local.txt
requests
boto3
fabric>=2.4.0,<3
gitpython>=2.1.11,<2.2
